+++ probe-rs/src/bin/probe-rs/util/logging.rs
@@ -4,15 +4,6 @@ use is_terminal::IsTerminal;
 use log::{Level, LevelFilter, Log, Record};
 use once_cell::sync::Lazy;
 use pretty_env_logger::env_logger::{Builder, Logger};
-#[cfg(feature = "sentry")]
-use sentry::{
-    integrations::panic::PanicIntegration,
-    types::{Dsn, Uuid},
-    Breadcrumb,
-};
-use simplelog::{CombinedLogger, SharedLogger};
-#[cfg(feature = "sentry")]
-use std::{borrow::Cow, panic::PanicInfo, str::FromStr};
 use std::{
     fmt::{self},
     io::Write,
@@ -21,16 +12,12 @@ use std::{
         Arc, RwLock,
     },
 };
-use terminal_size::{Height, Width};
 
 /// The maximum window width of the terminal, given in characters possible.
 static MAX_WINDOW_WIDTH: AtomicUsize = AtomicUsize::new(0);
 
 /// Stores the progress bar for the logging facility.
 static PROGRESS_BAR: Lazy<RwLock<Option<Arc<ProgressBar>>>> = Lazy::new(|| RwLock::new(None));
-
-#[cfg(feature = "sentry")]
-static LOG: Lazy<Arc<RwLock<Vec<Breadcrumb>>>> = Lazy::new(|| Arc::new(RwLock::new(vec![])));
 
 /// A structure to hold a string with a padding attached to the start of it.
 struct Padded<T> {
@@ -66,12 +53,13 @@ fn colored_level(level: Level) -> ColoredString {
     }
 }
 
-struct ShareableLogger {
+/// Logger wrapper that can coexist peacefully with indicatif progressbars.
+struct CliLogger {
     env_logger: Logger,
     output_is_terminal: bool,
 }
 
-impl Log for ShareableLogger {
+impl Log for CliLogger {
     fn enabled(&self, metadata: &log::Metadata<'_>) -> bool {
         metadata.level() <= self.env_logger.filter()
     }
@@ -108,20 +96,6 @@ impl Log for ShareableLogger {
 
     fn flush(&self) {
         self.env_logger.flush();
-    }
-}
-
-impl SharedLogger for ShareableLogger {
-    fn level(&self) -> LevelFilter {
-        self.env_logger.filter()
-    }
-
-    fn config(&self) -> Option<&simplelog::Config> {
-        None
-    }
-
-    fn as_log(self: Box<Self>) -> Box<dyn log::Log> {
-        Box::new(*self)
     }
 }
 
@@ -168,51 +142,14 @@ pub fn init(level: Option<Level>) {
         writeln!(f, "       {} {} > {}", level, target, record.args())
     });
 
-    // Sentry logging (all log levels except tracing (to not clog the server disk & internet sink)).
-    #[cfg(feature = "sentry")]
-    let mut sentry = {
-        let mut sentry = Builder::new();
-
-        // Always use the Debug log level.
-        sentry.filter_level(LevelFilter::Debug);
-
-        // Define our custom log format.
-        sentry.format(move |_f, record| {
-            let mut log_guard = LOG.write().unwrap();
-            log_guard.push(Breadcrumb {
-                level: match record.level() {
-                    Level::Error => sentry::Level::Error,
-                    Level::Warn => sentry::Level::Warning,
-                    Level::Info => sentry::Level::Info,
-                    Level::Debug => sentry::Level::Debug,
-                    // This mapping is intended as unfortunately, Sentry does not have any trace level for events & breadcrumbs.
-                    Level::Trace => sentry::Level::Debug,
-                },
-                category: Some(record.target().to_string()),
-                message: Some(format!("{}", record.args())),
-                ..Default::default()
-            });
-
-            Ok(())
-        });
-
-        sentry
-    };
-
     let output_is_terminal = std::io::stderr().is_terminal();
 
-    CombinedLogger::init(vec![
-        Box::new(ShareableLogger {
-            env_logger: log_builder.build(),
-            output_is_terminal,
-        }),
-        #[cfg(feature = "sentry")]
-        Box::new(ShareableLogger {
-            env_logger: sentry.build(),
-            output_is_terminal,
-        }),
-    ])
-    .unwrap();
+    let logger = Box::new(CliLogger {
+        env_logger: log_builder.build(),
+        output_is_terminal,
+    });
+    log::set_max_level(logger.env_logger.filter());
+    log::set_boxed_logger(logger).unwrap();
 }
 
 /// Sets the currently displayed progress bar of the CLI.
@@ -227,34 +164,6 @@ pub fn clear_progress_bar() {
     *guard = None;
 }
 
-#[cfg(feature = "sentry")]
-fn send_logs() {
-    let mut log_guard = LOG.write().unwrap();
-
-    for breadcrumb in log_guard.drain(..) {
-        sentry::add_breadcrumb(breadcrumb);
-    }
-}
-
-#[cfg(feature = "sentry")]
-fn sentry_config(release: String) -> sentry::ClientOptions {
-    sentry::ClientOptions {
-        dsn: Some(
-            Dsn::from_str(
-                "https://820ae3cb7b524b59af68d652aeb8ac3a@o473674.ingest.sentry.io/5508777",
-            )
-            .unwrap(),
-        ),
-        release: Some(Cow::<'static>::Owned(release)),
-        #[cfg(debug_assertions)]
-        environment: Some(Cow::Borrowed("Development")),
-        #[cfg(not(debug_assertions))]
-        environment: Some(Cow::Borrowed("Production")),
-        default_integrations: false,
-        ..Default::default()
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct Metadata {
     pub chip: Option<String>,
@@ -262,154 +171,6 @@ pub struct Metadata {
     pub speed: Option<String>,
     pub release: String,
     pub commit: String,
-}
-
-#[cfg(feature = "sentry")]
-/// Sets the metadata concerning the current probe-rs session on the sentry scope.
-fn set_metadata(metadata: &Metadata) {
-    sentry::configure_scope(|scope| {
-        if let Some(chip) = metadata.chip.as_ref() {
-            scope.set_tag("chip", chip);
-        }
-        if let Some(probe) = metadata.probe.as_ref() {
-            scope.set_tag("probe", probe);
-        }
-        if let Some(speed) = metadata.speed.as_ref() {
-            scope.set_tag("speed", speed);
-        }
-        scope.set_tag("commit", &metadata.commit);
-    })
-}
-
-#[cfg(feature = "sentry")]
-const SENTRY_SUCCESS: &str = r"Your error was reported successfully. If you don't mind, please open an issue on Github and include the UUID:";
-
-#[cfg(feature = "sentry")]
-fn print_uuid(uuid: Uuid) {
-    let size = terminal_size::terminal_size();
-    if let Some((Width(w), Height(_h))) = size {
-        let lines = chunk_string(&format!("{SENTRY_SUCCESS} {uuid}"), w as usize - 14);
-
-        for (i, l) in lines.iter().enumerate() {
-            if i == 0 {
-                println!("  {} {}", "Thank You!".cyan().bold(), l);
-            } else {
-                println!("             {l}");
-            }
-        }
-    } else {
-        print!("{SENTRY_HINT}");
-    }
-}
-
-#[cfg(feature = "sentry")]
-/// Captures an anyhow error with sentry and sends all previously captured logs.
-pub fn capture_anyhow(metadata: &Metadata, error: &anyhow::Error) {
-    let _guard = sentry::init(sentry_config(metadata.release.clone()));
-    set_metadata(metadata);
-    send_logs();
-    let uuid = sentry::integrations::anyhow::capture_anyhow(error);
-    print_uuid(uuid);
-}
-
-#[cfg(feature = "sentry")]
-/// Captures a panic with sentry and sends all previously captured logs.
-pub fn capture_panic(metadata: &Metadata, info: &PanicInfo<'_>) {
-    let _guard = sentry::init(sentry_config(metadata.release.clone()));
-    set_metadata(metadata);
-    send_logs();
-    let uuid = sentry::capture_event(PanicIntegration::new().event_from_panic_info(info));
-    print_uuid(uuid);
-}
-
-/// Ask for a line of text.
-fn text() -> std::io::Result<String> {
-    // Read up to the first newline or EOF.
-
-    let mut out = String::new();
-    std::io::stdin().read_line(&mut out)?;
-
-    // Only capture up to the first newline.
-    if let Some(mut newline) = out.find('\n') {
-        if newline > 0 && out.as_bytes()[newline - 1] == b'\r' {
-            newline -= 1;
-        }
-        out.truncate(newline);
-    }
-
-    Ok(out)
-}
-
-const SENTRY_HINT: &str = r"Unfortunately probe-rs encountered an unhandled problem. To help the devs, you can automatically log the error to sentry.io. Your data will be transmitted completely anonymously and cannot be associated with you directly. To hide this message in the future, please set $PROBE_RS_SENTRY to 'true' or 'false'. Do you wish to transmit the data? Y/n: ";
-
-/// Chunks the given string into pieces of maximum_length whilst honoring word boundaries.
-fn chunk_string(s: &str, max_width: usize) -> Vec<String> {
-    let string = s.chars().collect::<Vec<char>>();
-
-    let mut result = vec![];
-
-    let mut last_ws = 0;
-    let mut offset = 0;
-    let mut i = 0;
-    let mut t_max_width = max_width;
-    while i < string.len() {
-        let c = string[i];
-        if c.is_whitespace() {
-            last_ws = i;
-        }
-        if i > offset + t_max_width {
-            if last_ws > offset {
-                let s = string[offset..last_ws].iter().collect::<String>();
-                result.push(s);
-                t_max_width = max_width;
-            } else {
-                t_max_width += 1;
-            }
-
-            offset = last_ws + 1;
-            i = last_ws + 1;
-        } else {
-            i += 1;
-        }
-    }
-    result.push(string[offset..].iter().collect::<String>());
-    result
-}
-
-/// Displays the text to ask if the crash should be reported.
-pub fn ask_to_log_crash() -> bool {
-    if let Ok(var) = std::env::var("PROBE_RS_SENTRY") {
-        var == "true"
-    } else {
-        let size = terminal_size::terminal_size();
-        if let Some((Width(w), Height(_h))) = size {
-            let lines = chunk_string(SENTRY_HINT, w as usize - 14);
-
-            for (i, l) in lines.iter().enumerate() {
-                if i == 0 {
-                    println!("        {} {}", "Hint".blue().bold(), l);
-                } else if i == lines.len() - 1 {
-                    print!("             {l}");
-                } else {
-                    println!("             {l}");
-                }
-            }
-        } else {
-            print!("{SENTRY_HINT}");
-        }
-
-        std::io::stdout().flush().ok();
-        let result = if let Ok(s) = text() {
-            let s = s.to_lowercase();
-            "yes".starts_with(&s)
-        } else {
-            false
-        };
-
-        println!();
-
-        result
-    }
 }
 
 /// Writes an error to stderr.