+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for propagating `CoreStatus` to the probe in use (#1588).
 - Added PY32F0xx series targets (#1619).
 
+### Removed
+
+- Removed Sentry integration in the CLI tools.
+
 ## [0.18.0]
 
 Released 2023-03-31